requirements.txt
lightning~=2.4.0
torch~=2.5.1
transformers~=4.46.3
pandas~=2.2.3
tensorboard~=2.18.0
scikit-learn~=1.5.2
torchmetrics~=1.6.0
PyJWT~=2.10.1
uvicorn~=0.34.0
fastapi~=0.115.6
pydantic~=2.10.4